==> 3. EPC_ETL_3_User_items.ipynb <==
# **EDUARDO PEREZ CHAVARRIA. FT17**
## Proyecto individual
### ETL 3. Archivo a analizar User_items

## **Cargar base**

import pandas as pd
import gzip
import ast

# Ruta al archivo comprimido gzip
archivo_gzip = "bases/users_items.json.gz"

# Lista para almacenar las filas del DataFrame
filas_usitems = []

# Abrir el archivo comprimido gzip y leer línea por línea
with gzip.open(archivo_gzip, 'rt', encoding='utf-8') as f:
    for line in f:
        # Evaluar cada línea y agregarla a la lista
        filas_usitems.append(ast.literal_eval(line))

# Crear un DataFrame a partir de la lista de filas
df_usit = pd.DataFrame(filas_usitems)

# Mostrar el DataFrame
print(df_usit)


Se elimina la columna user_url e items_count

df_usit = df_usit.drop(["user_url", "items_count"], axis=1)

df_usit

Explorar valor nulo

# Ver el tipo de datos de df_usit
tipos_datos = df_usit.dtypes

# Verificar la presencia de valores nulos
valores_nulos = df_usit.isnull().sum()

# Crear un nuevo DataFrame con la información
df_info = pd.DataFrame({
    "Tipos de Datos": tipos_datos,
    "Valores Nulos": valores_nulos
})

# Mostrar el DataFrame resultante
print(df_info)


ver columna items

df_usit["items"][0]

desanidar

# se normaliza la columna "items"
df_usit2 = pd.json_normalize(filas_usitems, record_path=["items"], meta=["steam_id","user_id"] )
df_usit=df_usit2


df_usit

# Definir el nuevo orden de las columnas
nuevo_orden = ["user_id", "steam_id", "item_id", "item_name", "playtime_forever", "playtime_2weeks"]

# Reorganizar las columnas del DataFrame
df_usit = df_usit[nuevo_orden]

# Mostrar el DataFrame resultante
df_usit


# Verificar si hay valores nulos en todas las columnas
nulos_por_columna = df_usit.isnull().sum()

# Mostrar la frecuencia de nulos por columna
print("Frecuencia de valores nulos por columna:")
print(nulos_por_columna)


# Verificar duplicados en df_usit
duplicados = df_usit.duplicated()

# Contar duplicados
cantidad_duplicados = duplicados.sum()


# Mostrar resultados
print("Cantidad total de registros duplicados en 'df_usit':", cantidad_duplicados)


eliminamos esos duplicados

df_usit.drop_duplicates(inplace=True)

# verificamos de nuevo,
duplicados = df_usit.duplicated()
# Contar duplicados
cantidad_duplicados = duplicados.sum()
# Mostrar resultados
print("Cantidad total de registros duplicados en 'df_usit':", cantidad_duplicados)


la columna playtime_2weeks no nos interesa porque la info sobre el tiempo de juego total está en play_time_forever. La quitamos

# Dropear la columna "playtime_2weeks"
df_usit = df_usit.drop(columns=["playtime_2weeks"])

# Verificar que la columna fue eliminada
df_usit

revisamos cómo ha quedado el df

# Frecuencia de valores únicos en cada columna
for columna in df_usit.columns:
    print(f"Frecuencia de valores únicos en la columna '{columna}':")
    print(df_usit[columna].value_counts())
    print("\n")


print(df_usit.count())


Guardamos el df final

#data_curado = "bases/users_items_curado.csv"
#df_usit.to_csv(data_curado, index=False, encoding="utf-8")
#print(f'Has guardado tu base:  {data_curado}')

df_usit.to_parquet("bases/user_items_curado.parquet")

Importamos

ruta_archivo_parquet = "bases/user_items_curado.parquet"

# Importar el archivo Parquet como un DataFrame
df_parquet = pd.read_parquet(ruta_archivo_parquet)

Checamos que esté bien

comprobamos que no se nos ha ido a perdido algun valor

df_parquet.isnull().sum()

df_parquet
