==> 4. EPC_FE_analisis_sentimientos.ipynb <==
# **EDUARDO PEREZ CHAVARRIA. FT17**
## Proyecto individual
### FE. Análisis de sentimientos



import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import csv

import re
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from textblob import TextBlob

from langdetect import detect

## **1. Analisis de sentimientos**

### **Importamos el archivo ya curado que contiene las reviews de usuarios**

data_curado = "bases/users_review_curado.csv"

# Importar el archivo CSV como DataFrame
df_usrev = pd.read_csv(data_curado, encoding="utf-8")

# Mostrar información del DataFrame importado
print(df_usrev.info())


### **Realizamos el análisis de sentimientos**

# Descarga los recursos necesarios de NLTK
nltk.download("vader_lexicon")

# Inicializa el analizador de sentimientos de NLTK
sia = SentimentIntensityAnalyzer()

# Función para asignar valores según la escala
def obtener_puntaje_sentimiento(texto):
    if pd.isnull(texto) or texto == "":
        return 1  # Retorna neutral si está vacío o es NaN
    elif isinstance(texto, str):
        # Analiza el sentimiento usando Vader de NLTK
        vader_sentimiento = sia.polarity_scores(texto)
        vader_puntaje_compuesto = vader_sentimiento["compound"]

        # Analiza el sentimiento usando TextBlob
        textblob_analysis = TextBlob(texto)
        polarity = textblob_analysis.sentiment.polarity

        # Aplica la escala para determinar el puntaje final
        if vader_puntaje_compuesto >= -0.05 and polarity >= -0.2 and polarity <= 0.2:
            return 1  # Neutral
        elif vader_puntaje_compuesto >= -0.05 or polarity > 0.2:
            return 2  # Buen puntaje
        elif vader_puntaje_compuesto <= -0.05 or polarity < -0.2:
            return 0  # Mal puntaje
    return 1  # Retorna neutral para otros casos

# Convierte la columna "review" a str si no lo es
df_usrev["review"] = df_usrev["review"].astype(str)

# Aplica la función obtener_puntaje_sentimiento a la columna "review"
df_usrev["Sentiment_Score"] = df_usrev["review"].apply(obtener_puntaje_sentimiento)

# Visualiza el DataFrame
print(df_usrev)


### **Explorar en qué idioma están las reseña**

Vamos a explorar en qué idioma están las reseñas para hacernos una idea de la composición de la columna. En teoría vader funciona bien para cualquier idioma aunque fue entrenado principalmente con texto en inglés.

# Función para detectar el idioma de la reseña
def detectar_idioma(texto):
    try:
        return detect(texto)
    except:
        return "No detectado"  # Si no se puede detectar el idioma, se marca como "No detectado"

# Aplicar la función a la columna "review" y crear una nueva columna "idioma"
df_usrev["idioma"] = df_usrev["review"].apply(detectar_idioma)

df_usrev

### **Valores unicos para los idiomas**

conteo_idiomas = df_usrev["idioma"].value_counts()
print(conteo_idiomas)


Notamos que casi el 80% de las reseñas están en inglés. Por lo que el análisis sigue siendo bastante valido. Vamos a explorar los casos donde no se detectó el idioma y aun así realizó análisis de sentimientos

# Filtra las filas donde el idioma es "No detectado" y muestra las filas relevantes
sin_idioma_detectado = df_usrev[df_usrev["idioma"] == "No detectado"]
resultados_analisis = sin_idioma_detectado[["user_id", "item_id", "recommend", "review", "year", "Sentiment_Score", "idioma"]]

# Visualiza los análisis de sentimiento para las reseñas sin idioma detectado
print(resultados_analisis)

# Exporta los resultados a un archivo de texto
resultados_analisis.to_csv("resultados_analisis_sin_idioma_detectado.txt", sep="\t", index=False)


Se puede observar que varios de los casos contienen simbolos que pueden tomarse como evaluaciones positivas. Por ello decidimos conservarlos. 

df_usrev[["review", "Sentiment_Score"]]


### **Frencuencias de la codificacion de analisis**

frecuencias = df_usrev["Sentiment_Score"].value_counts()
print(frecuencias)


### **Reseñas no detectadas **

con el txt que exportamos antes podemos notar que hay varias respuestas tipo <3, corazones, 100/10, 😄, 9/11, ♥, :), 💋, < 3, ★★★★★ que indican que la reseña es positiva, vamos a cambiarlas cuando haya ese tipo de respuesta

def asignar_sentimiento_especial(texto):
    # Expresiones que indican un sentimiento positivo no neutro ni negativo
    expresiones_positivas = [r"< ?3{3,}", r"♥{3,}", r":\){3,}", r"★{3,}"]

    # Expresiones que indican una calificación alta
    expresiones_calificacion = [r"\d{3,5}/10", r"[8-9]/10|[1-9][0-9]{1,3}", r"[8-9]/[10-1000]"]

    for exp in expresiones_positivas + expresiones_calificacion:
        if re.search(exp, texto):
            return 2  # Sentimiento positivo alto

    return 1  # Si no se encuentra ninguna expresión especial, se asigna neutral

# Crear una máscara para las filas con "idioma" igual a "No detectado"
mask_no_detectado = df_usrev["idioma"] == "No detectado"

# Aplicar la función a las filas con "idioma" igual a "No detectado"
df_usrev.loc[mask_no_detectado, "Sentiment_Score"] = df_usrev.loc[mask_no_detectado, "review"].apply(asignar_sentimiento_especial)

# Visualizar los cambios en el DataFrame
print(df_usrev)

# Exportar a un archivo de texto solo las filas con "idioma" igual a "No detectado"
resultados_analisis = df_usrev[mask_no_detectado]
resultados_analisis.to_csv("resultados_analisis_sentimiento_especial.txt", sep="\t", index=False)


Vamos a ver cómo han cambiado los valores unicos. Antes del cambio eran estos

* Sentiment_Score
* 1    29735
* 2    19901
* 0     8764
* Name: count, dtype: int64**

frecuencias = df_usrev["Sentiment_Score"].value_counts()
print(frecuencias)

Se aprecian pocos cambios significativos al implementar esa medida de cuidado. Vamos a eliminar la columna de idioma que ya no utilizaremos

df_usrev.drop(columns=["idioma"], inplace=True)

## **Observamos el df final**

df_usrev

## **Guardamos la base**

data_curado_sentiments = "bases/users_review_curado_sentiments.parquet"

# Guardar el DataFrame en formato Parquet con comillas dobles
df_usrev.to_parquet(data_curado_sentiments, index=False, engine="pyarrow", compression="gzip")

print(f"Has guardado tu base en formato Parquet con comillas dobles: {data_curado_sentiments}")


# Cargar el archivo Parquet de nuevo
df_importado = pd.read_parquet("bases/users_review_curado_sentiments.parquet")

# Verificar que se haya cargado correctamente
df_importado
