==> 8. funcion item-similarity.ipynb <==
# **EPC. FUNCION MODELO DE RECOMENDACION ITEM SIMI**

import pandas as pd
import pyarrow.parquet as pq

ruta_archivo_item_simi = 'bases/item_simi_df.parquet'
df_item_simi = pd.read_parquet(ruta_archivo_item_simi)

# Importar la matriz de similitud desde el archivo Parquet
ruta_archivo_item_simi = 'bases/item_simi_df.parquet'
df_item_simi = pq.read_table(ruta_archivo_item_simi).to_pandas()



## 1. FUNCION

SIN GENERO

def recomendacion_juego(nombre_juego):
    
    # Obtener la fila correspondiente al juego ingresado
    juego_row = df_item_simi.loc[nombre_juego]

    # Ordenar los índices de los juegos según la similitud (de mayor a menor)
    juegos_similares_indices = juego_row.argsort()[::-1]

    # Seleccionar los primeros 5 juegos más similares (excluyendo el propio juego)
    top_5_juegos = juegos_similares_indices[1:6]

    # Obtener nombres de los juegos recomendados
    juegos_recomendados_dict = {f"Juego {i+1}": juego for i, juego in enumerate(df_item_simi.index[top_5_juegos])}

    return juegos_recomendados_dict



# Ejemplo de uso:
recomendaciones_juegos = recomendacion_juego("Killing Floor")

# Visualizar las recomendaciones

recomendaciones_juegos
